Changed from perl module to standalone script
Previously this had a perl module that implemented a class
(Doxygen::Lua) which was instantiated from a script and then called a
single function (parse). This commit leaves just a single script with
the parse function. This means installation is trivial -- nothing needs
to be installed into the perl module tree -- and the install is now done
using cmake. The reason for this is that we need to install additional
files for compatability with torch-pkg and torch-dok and this was
unnecessarily complicated using Module::Install.